Add files via upload
I expanded on the signup feature so when the user has an account they can just login. On top of that when the user has an account either from creating one or logging in the sign up/login button is removed and replaced with a user icon. Finally I implemented a drop down that appears when clicking on the user icon.